refactor: create sub-package for html parser
